fix(viewer): close M-06 telemetry and split large modules
