+++ tests/test_m06_single_rebase_contract.py
@@ -27,8 +27,8 @@ EXPECTED_ANCHOR_INVENTORY = {
     ("src/tiles3d/renderer.rs", "from_buildings", "rebase_call", 1),
     ("src/viewer/camera_controller.rs", "prospective_anchor", "anchor_copy", 1),
     ("src/viewer/camera_controller.rs", "prospective_anchor", "rebase_call", 1),
-    ("src/viewer/event_loop/command_batch.rs", "preflight_command_batch", "anchor_copy", 1),
-    ("src/viewer/event_loop/command_batch.rs", "preflight_command_batch", "rebase_call", 1),
+    ("src/viewer/event_loop/command_preflight.rs", "preflight_command_batch", "anchor_copy", 1),
+    ("src/viewer/event_loop/command_preflight.rs", "preflight_command_batch", "rebase_call", 1),
     ("src/viewer/init/viewer_new.rs", "new", "constructor", 1),
     ("src/viewer/input/viewer_input.rs", "handle_input", "anchor_copy", 1),
     ("src/viewer/input/viewer_input.rs", "handle_input", "anchor_copy", 2),
